sbsigntool: use custom build/configure
There is an ./autogen.sh script that need to be launched before running
./configure.

Currently, this version of sbsigntool (0.9.1) fails to compile - both
the host package and the target package.

The host failure is caused by duplication of include dirs, which results
in *redefinition*. Probably more than that.

Missing execinfo.h on musl probably the cause of failure on target,
although more build errors would be produced.

Signed-off-by: Alif M. Ahmad <alive4ever@live.com>